move nullstore out of CheckpointStorageProvider
